PR:	207068
Submitted by:	joss.upton@yahoo.com
MFC after:	2 weeks

+++ sys/dev/hwpmc/hwpmc_mod.c
@@ -1483,7 +1483,7 @@ pmc_process_csw_out(struct thread *td)
 				 * increasing monotonically, modulo a 64
 				 * bit wraparound.
 				 */
-				KASSERT((int64_t) tmp >= 0,
+				KASSERT(tmp >= 0,
 				    ("[pmc,%d] negative increment cpu=%d "
 				     "ri=%d newvalue=%jx saved=%jx "
 				     "incr=%jx", __LINE__, cpu, ri,
